Made batch files as makefile equivalents for Windows.
Also added an existence test into reset.empty() so it doesn't
try to list files in a non-existent dir. Apparently Win7
doesn't like to do that.

+++ reset.py
@@ -22,6 +22,8 @@ class Reset:
         if path == '':
             path = self.destDir
         path = self.filesystem.joinPath(path)
+        if not self.filesystem.checkDirExists(path):
+            return
         self.destImages = self.filesystem.listJpegs(path)
         for image in self.destImages:
             self.filesystem.removeFile([path, image])
